Move DeviceSelectionPopup to its right place
